Send photos after text not in the middle; Fix videos

+++ src/postSender/sendPost/index.ts
@@ -6,10 +6,10 @@ import hasBanWords from '../hasBanWords';
 import getLinksText from '../getLinksText';
 import parseAttachments from '../parseAttachments';
 import prepareText from '../prepareText';
-import { exec as ytdl } from 'youtube-dl-exec';
 import bot from '../../telegram';
 import config from '../../config';
 import fs from "fs"
+import spawn from 'await-spawn';
 
 
 async function sendPost(post) {
@@ -21,30 +21,35 @@ async function sendPost(post) {
 
   if (photos.length === 0) {
     if (text.length + linksText.length === 0) return;
-    sendText(text, linksText);
+    await sendText(text, linksText);
   } else if (photos.length === 1) {
-    sendPhoto(photos[0], text, linksText);
+    await sendPhoto(photos[0], text, linksText);
   } else {
-    sendPhotos(photos, text, linksText);
+    await sendPhotos(photos, text, linksText);
   }
 
   if (videos.length !== 0) {
+    console.log("Downloading videos")
     for (const video of videos) {
       const channel = config.get('channel');
+      if(video.url.includes("youtube") || video.url.includes("youtu.be")) {
+        console.log(`Sending video ${video.url}`)
+        await bot.sendMessage(channel, video.url)
+        continue
+      }
+      console.log(`Downloading video ${video.url}`)
       try {
-        const subprocess = ytdl(video.url)
+        const binary = config.get("youtube-dl-binary")
         const filename = `${new Date().getTime()}.tmp`
-        subprocess.stdout.pipe(fs.createWriteStream(filename))
-        subprocess.on("exit", (code) => {
-          if (code !== 0) {
-            bot.sendMessage(channel, `Couldn't download video ${video.url}`).catch(console.error)
-            console.error(subprocess.stderr)
-          } else {
-            bot.sendVideo(channel, filename).catch(console.error)
-          }
-          fs.unlink(filename, (err) => {
-            if (err !== null) console.error(err)
-          })
+        console.log(`Downloading video to ${filename} with binary ${binary}`)
+        await spawn(binary, ["-o", filename, video.url], {
+          "stdio": "inherit"
+        })
+        console.log("Downloaded video from VK")
+        await bot.sendVideo(channel, filename)
+        console.log("Uploaded to Telegram")
+        fs.unlink(filename, (err) => {
+          if (err !== null) console.error(err); else console.log("Deleted temporary file")
         })
       } catch (e) {
         await bot.sendMessage(channel, `Couldn't download video ${video.url}`)
